+++ sources/librovore/__/imports.py
@@ -25,6 +25,7 @@
 
 import                      abc
 import                      asyncio
+import                      base64
 import collections.abc as   cabc
 import                      collections
 import contextlib as        ctxl
@@ -36,6 +37,7 @@ import                      io
 import                      inspect
 import                      json
 import                      locale
+import                      os
 import                      platform
 import                      re
 import                      shutil